- classes  modelling the "entity" layer of the communication between DECSIRA and IRIDE service, being common between the sec-iride plugin and the custom Resource Access Manager, have been refactored and moved into the separate, new, module iride-entities, to have them packaged in a separate Jar, as an independent artifact
