+++ security/pom.xml
@@ -13,8 +13,7 @@
 
     <!-- sotto moduli del layer di sicurezza -->
     <modules>
-        <!-- temporary commented -->
-        <!-- module>iride-entities</module -->
+        <module>iride-entities</module>
         <module>sec-iride</module>
     </modules>
