+++ security/sec-iride/pom.xml
@@ -137,6 +137,13 @@
     </repositories>
 
     <dependencies>
+        <!-- sira modules -->
+        <dependency>
+		    <groupId>org.geoserver.security</groupId>
+		    <artifactId>iride-entities</artifactId>
+		    <version>1.0-SNAPSHOT</version>
+        </dependency>
+
         <!-- GeoServer -->
         <dependency>
             <groupId>org.geoserver</groupId>
