--- security/sec-iride/src/main/java/org/geoserver/security/iride/entity/exception/IrideEntitySerializationException.java
@@ -1,51 +1,0 @@
-/*
- *  GeoServer Security Provider plugin used for doing authentication and authorization operations using CSI-Piemonte IRIDE Service.
- *  Copyright (C) 2016  Regione Piemonte (www.regione.piemonte.it)
- *
- *  This program is free software; you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation; either version 2 of the License, or
- *  (at your option) any later version.
- *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License along
- *  with this program; if not, write to the Free Software Foundation, Inc.,
- *  51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
- */
-package org.geoserver.security.iride.entity.exception;
-
-import org.geoserver.security.iride.exception.IrideSecuritySystemException;
-
-/**
- * <code>IRIDE</code> entity object serialization exception class.
- *
- * @author "Simone Cornacchia - seancrow76@gmail.com, simone.cornacchia@consulenti.csi.it (CSI:71740)"
- */
-public final class IrideEntitySerializationException extends IrideSecuritySystemException {
-
-    private static final long serialVersionUID = 4001069701076979150L;
-
-    /**
-     * Constructor.
-     *
-     * @param message
-     */
-    public IrideEntitySerializationException(String message) {
-        super(message);
-    }
-
-    /**
-     * Constructor.
-     *
-     * @param message
-     * @param cause
-     */
-    public IrideEntitySerializationException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-}
