+++ security/sec-iride/src/test/java/org/geoserver/security/iride/AllTests.java
@@ -18,7 +18,6 @@
  */
 package org.geoserver.security.iride;
 
-import org.geoserver.security.iride.util.IrideIdentityValidityTest;
 import org.geoserver.security.iride.util.factory.template.freemarker.FreeMarkerConfigurationDefaultFactoryTest;
 import org.geoserver.security.iride.util.factory.template.freemarker.FreeMarkerTemplateEngineFactoryTest;
 import org.geoserver.security.iride.util.factory.template.freemarker.roleservice.IrideRoleServiceFactoryTest;
@@ -35,12 +34,10 @@ import org.junit.runners.Suite.SuiteClasses;
 @SuiteClasses({
     IrideRoleServiceTest.class,
     IrideUserGroupServiceTest.class,
-    IrideIdentityValidityTest.class,
     IrideRoleServiceFactoryTest.class,
     FreeMarkerConfigurationDefaultFactoryTest.class,
     FreeMarkerTemplateEngineFactoryTest.class,
     org.geoserver.security.iride.entity.converter.AllTests.class,
-    org.geoserver.security.iride.identity.AllTests.class,
     org.geoserver.security.iride.service.policy.handler.request.AllTests.class,
     org.geoserver.security.iride.soap.request.iride.AllTests.class,
     org.geoserver.security.iride.util.template.AllTests.class,
